feat: Use store
