+++ src/main.js
@@ -1,10 +1,13 @@
 import "./assets/main.css";
 
 import { createApp } from "vue";
+import { createPinia } from "pinia";
 import App from "./App.vue";
 
 const app = createApp(App);
+const pinia = createPinia();
 
 app.use(app);
+app.use(pinia);
 
 app.mount("#app");
